+++ pkg/scalers/aws_sqs_queue_scaler_test.go
@@ -3,6 +3,7 @@ package scalers
 import (
 	"context"
 	"errors"
+	"strconv"
 	"testing"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -25,6 +26,10 @@ const (
 
 	testAWSSQSErrorQueueURL   = "https://sqs.eu-west-1.amazonaws.com/account_id/Error"
 	testAWSSQSBadDataQueueURL = "https://sqs.eu-west-1.amazonaws.com/account_id/BadData"
+
+	testAWSSQSApproximateNumberOfMessagesVisible    = 200
+	testAWSSQSApproximateNumberOfMessagesNotVisible = 100
+	testAWSSQSApproximateNumberOfMessagesDelayed    = 50
 )
 
 var testAWSSQSEmptyResolvedEnv = map[string]string{}
@@ -65,14 +70,16 @@ func (m *mockSqs) GetQueueAttributes(input *sqs.GetQueueAttributesInput) (*sqs.G
 			Attributes: map[string]*string{
 				"ApproximateNumberOfMessages":           aws.String("NotInt"),
 				"ApproximateNumberOfMessagesNotVisible": aws.String("NotInt"),
+				"ApproximateNumberOfMessagesDelayed":    aws.String("NotInt"),
 			},
 		}, nil
 	}
 
 	return &sqs.GetQueueAttributesOutput{
 		Attributes: map[string]*string{
-			"ApproximateNumberOfMessages":           aws.String("200"),
-			"ApproximateNumberOfMessagesNotVisible": aws.String("100"),
+			"ApproximateNumberOfMessages":           aws.String(strconv.Itoa(testAWSSQSApproximateNumberOfMessagesVisible)),
+			"ApproximateNumberOfMessagesNotVisible": aws.String(strconv.Itoa(testAWSSQSApproximateNumberOfMessagesNotVisible)),
+			"ApproximateNumberOfMessagesDelayed":    aws.String(strconv.Itoa(testAWSSQSApproximateNumberOfMessagesDelayed)),
 		},
 	}, nil
 }
@@ -327,6 +334,44 @@ var awsSQSGetMetricTestData = []*parseAWSSQSMetadataTestData{
 		false,
 		"not error with scaleOnInFlight enabled"},
 	{map[string]string{
+		"queueURL":       testAWSSQSProperQueueURL,
+		"queueLength":    "1",
+		"awsRegion":      "eu-west-1",
+		"scaleOnDelayed": "false"},
+		testAWSSQSAuthentication,
+		testAWSSQSEmptyResolvedEnv,
+		false,
+		"not error with scaleOnDelayed disabled"},
+	{map[string]string{
+		"queueURL":       testAWSSQSProperQueueURL,
+		"queueLength":    "1",
+		"awsRegion":      "eu-west-1",
+		"scaleOnDelayed": "true"},
+		testAWSSQSAuthentication,
+		testAWSSQSEmptyResolvedEnv,
+		false,
+		"not error with scaleOnDelayed enabled"},
+	{map[string]string{
+		"queueURL":        testAWSSQSProperQueueURL,
+		"queueLength":     "1",
+		"awsRegion":       "eu-west-1",
+		"scaleOnInFlight": "false",
+		"scaleOnDelayed":  "false"},
+		testAWSSQSAuthentication,
+		testAWSSQSEmptyResolvedEnv,
+		false,
+		"not error with scaledOnInFlight and scaleOnDelayed disabled"},
+	{map[string]string{
+		"queueURL":        testAWSSQSProperQueueURL,
+		"queueLength":     "1",
+		"awsRegion":       "eu-west-1",
+		"scaleOnInFlight": "true",
+		"scaleOnDelayed":  "true"},
+		testAWSSQSAuthentication,
+		testAWSSQSEmptyResolvedEnv,
+		false,
+		"not error with scaledOnInFlight and scaleOnDelayed enabled"},
+	{map[string]string{
 		"queueURL":        testAWSSQSErrorQueueURL,
 		"queueLength":     "1",
 		"awsRegion":       "eu-west-1",
@@ -390,11 +435,17 @@ func TestAWSSQSScalerGetMetrics(t *testing.T) {
 		case testAWSSQSBadDataQueueURL:
 			assert.Error(t, err, "expect error because of bad data return from sqs")
 		default:
+			expectedMessages := testAWSSQSApproximateNumberOfMessagesVisible
+
 			if meta.scaleOnInFlight {
-				assert.EqualValues(t, int64(300.0), value[0].Value.Value())
-			} else {
-				assert.EqualValues(t, int64(200.0), value[0].Value.Value())
+				expectedMessages += testAWSSQSApproximateNumberOfMessagesNotVisible
 			}
+
+			if meta.scaleOnDelayed {
+				expectedMessages += testAWSSQSApproximateNumberOfMessagesDelayed
+			}
+
+			assert.EqualValues(t, int64(expectedMessages), value[0].Value.Value())
 		}
 	}
 }